Moved files around
Enact-DCO-1.0-Signed-off-by: Derek Tor <derek.tor@lge.com>
