+++ packages/core/internal/Registry/Registry.js
