+++ packages/core/internal/Registry/package.json
